requirements.txt
configparser==3.5.0
Cython==0.26
dbus-python==1.2.4
enum34==1.1.6
flake8==3.4.1
funcsigs==1.0.2
jedi==0.10.2
mccabe==0.6.1
mock==2.0.0
numpy==1.13.1
pbr==3.1.1
pep8==1.7.0
psutil==5.2.2
pycodestyle==2.3.1
pyflakes==1.5.0
pyglet==1.2.4
pyreadline==2.1
pyserial==3.4
six==1.11.0
